Nalu Kernels SIMD vectorization WIP
    - Convert ScratchViews and MasterElementViews to templated classes
      so that they can be used both with `double` and `DoubleType`

    - Modify related functions to support template parameters.
