+++ include/MomentumBuoyancySrcElemKernel.h
@@ -22,7 +22,6 @@ namespace nalu {
 class SolutionOptions;
 class MasterElement;
 class ElemDataRequests;
-class ScratchViews;
 
 /** CMM buoyancy term for momentum equation (velocity DOF)
  */
@@ -43,7 +42,7 @@ public:
   virtual void execute(
     SharedMemView<DoubleType**>&,
     SharedMemView<DoubleType*>&,
-    ScratchViews&);
+    ScratchViews<DoubleType>&);
 
 private:
   MomentumBuoyancySrcElemKernel() = delete;
