+++ include/ScalarDiffElemKernel.h
@@ -22,7 +22,6 @@ namespace nalu {
 class SolutionOptions;
 class MasterElement;
 class ElemDataRequests;
-class ScratchViews;
 
 /** CVFEM scalar advection/diffusion kernel
  */
@@ -45,7 +44,7 @@ public:
   virtual void execute(
     SharedMemView<DoubleType**>&,
     SharedMemView<DoubleType*>&,
-    ScratchViews&);
+    ScratchViews<DoubleType>&);
 
 private:
   ScalarDiffElemKernel() = delete;
