+++ include/ScalarMassElemKernel.h
@@ -23,7 +23,6 @@ class TimeIntegrator;
 class SolutionOptions;
 class MasterElement;
 class ElemDataRequests;
-class ScratchViews;
 
 /** CMM (BDF2/BE) for scalar equation
  */
@@ -50,7 +49,7 @@ public:
   virtual void execute(
     SharedMemView<DoubleType**>&,
     SharedMemView<DoubleType*>&,
-    ScratchViews&);
+    ScratchViews<DoubleType>&);
 
 private:
   ScalarMassElemKernel() = delete;
