+++ include/SupplementalAlgorithm.h
@@ -20,7 +20,6 @@ namespace sierra{
 namespace nalu{
 
 class Realm;
-class ScratchViews;
 
 class SupplementalAlgorithm
 {
@@ -32,12 +31,6 @@ public:
   virtual ~SupplementalAlgorithm() {}
 
   virtual void setup() {}
-
-  virtual void element_execute(
-    SharedMemView<double**>& lhs,
-    SharedMemView<double*>& rhs,
-    stk::mesh::Entity element,
-    ScratchViews& scratchViews) {}
 
   virtual void elem_execute(
     double *lhs,
