+++ include/nso/MomentumNSOKeElemKernel.h
@@ -22,7 +22,6 @@ namespace nalu {
 class SolutionOptions;
 class MasterElement;
 class ElemDataRequests;
-class ScratchViews;
 
 /** NSO for momentum equation
  *
@@ -44,7 +43,7 @@ public:
   virtual void execute(
     SharedMemView<DoubleType**>&,
     SharedMemView<DoubleType*>&,
-    ScratchViews&);
+    ScratchViews<DoubleType>&);
 
 private:
   MomentumNSOKeElemKernel() = delete;
