+++ include/nso/MomentumNSOSijElemKernel.h
@@ -22,7 +22,6 @@ namespace nalu {
 class SolutionOptions;
 class MasterElement;
 class ElemDataRequests;
-class ScratchViews;
 
 /** NSO for momentum equation
  *
@@ -42,7 +41,7 @@ public:
   virtual void execute(
     SharedMemView<DoubleType**>&,
     SharedMemView<DoubleType*>&,
-    ScratchViews&);
+    ScratchViews<DoubleType>&);
 
 private:
   MomentumNSOSijElemKernel() = delete;
