+++ src/ContinuityMassElemKernel.C
@@ -86,7 +86,7 @@ void
 ContinuityMassElemKernel<AlgTraits>::execute(
   SharedMemView<DoubleType **>&/*lhs*/,
   SharedMemView<DoubleType *>&rhs,
-  ScratchViews& scratchViews)
+  ScratchViews<DoubleType>& scratchViews)
 {
   const DoubleType projTimeScale = dt_/gamma1_;
