+++ src/nso/MomentumNSOElemKernel.C
@@ -126,7 +126,7 @@ void
 MomentumNSOElemKernel<AlgTraits>::execute(
   SharedMemView<DoubleType**>& lhs,
   SharedMemView<DoubleType *>& rhs,
-  ScratchViews& scratchViews)
+  ScratchViews<DoubleType>& scratchViews)
 {
   DoubleType w_dukdxScs   [AlgTraits::nDim_];
   DoubleType w_rhoVrtmScs [AlgTraits::nDim_];
