+++ src/nso/MomentumNSOSijElemKernel.C
@@ -86,7 +86,7 @@ void
 MomentumNSOSijElemKernel<AlgTraits>::execute(
   SharedMemView<DoubleType**>& lhs,
   SharedMemView<DoubleType *>& rhs,
-  ScratchViews& scratchViews)
+  ScratchViews<DoubleType>& scratchViews)
 {
   DoubleType w_ke         [AlgTraits::nodesPerElement_];
   DoubleType w_rhoVrtmScs [AlgTraits::nDim_];
